Fluff C3 seam 6: extract find / redline-resolution / clipboard hooks
editorHooks.ts carries three verbatim-moved behaviour hooks:
useFindInDocument (query state, decoration dispatch, active-match
scroll, Cmd-F/Cmd-S shortcuts), useTrackedChangeResolution (redlines
visibility, anchored count, per-edit and accept/reject-all resolution
against the real endpoints, widget handler wiring) and
useClipboardActions (copy/download/link/image-URL). Effect bodies,
dep arrays and the deliberately dep-less shortcut listener are
unchanged; the component passes in editor/plainText/canSave/save.
